657: Replace hard-coded values with constants and split code into files
